requirements.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==3.7.1
bcrypt==4.1.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==41.0.7
distlib==0.3.7
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.1.0.post1
fastapi==0.104.1
filelock==3.13.1
fsspec==2023.12.2
gradio_client==0.2.7
greenlet==3.0.1
gusml==1.0
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.25.2
huggingface-hub==0.19.4
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.3.0
MarkupSafe==2.1.3
mypy==1.7.1
mypy-extensions==1.0.0
orjson==3.9.10
packaging==23.2
passlib==1.7.4
platformdirs==4.1.0
psycopg==3.1.14
psycopg-binary==3.1.14
psycopg-pool==3.2.0
psycopg2 @ git+https://github.com/nwcell/psycopg2-windows.git@4039d982152492453abdfb72d0054230fcb42c30
psycopg2-binary==2.9.9
pyasn1==0.5.1
pycparser==2.21
pydantic==2.5.2
pydantic-extra-types==2.1.0
pydantic-settings==2.1.0
pydantic_core==2.14.5
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.0
starlette==0.27.0
tqdm==4.66.1
typer==0.9.0
types-passlib==1.7.7.13
types-psycopg2==2.9.21.19
types-pyasn1==0.5.0.1
types-python-jose==3.3.4.8
typing_extensions==4.8.0
tzdata==2023.3
ujson==5.8.0
urllib3==2.1.0
uvicorn==0.24.0.post1
watchfiles==0.21.0
websockets==12.0
